api/apikite: rename socialkite -> apikite
