+++ go/src/koding/api/apikite/socialkite.go
@@ -1,4 +1,4 @@
-package socialkite
+package apikite
 
 import (
 	"sync"
